+++ packages/ui/src/pages/HomePage/useRemoveAuthParams/constants.ts
@@ -1,0 +1,1 @@
+export const AUTH_PARAMS = ['code', 'state', 'scope'];
