+++ packages/ui/src/pages/HomePage/useRemoveAuthParams/index.ts
@@ -1,0 +1,1 @@
+export { default } from './useRemoveAuthUrlParams';
